borrado de dto, entity, servicios y repositorios de articuloManufacturado, añadir campo tiempococina en articuloConsumo
