--- src/main/java/com/example/demo/dtos/ArticuloManufacturadoDto.java
@@ -1,114 +1,0 @@
-
-package com.example.demo.dtos;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class ArticuloManufacturadoDto  {
-
-	
-
-	
-	private long id;
-	
-	
-	private int tiempoEstimadoCocina;
-	
-	
-	private String denominacion;
-	
-	
-	private double precioVenta;
-	
-	private String imagen;
-	
-	
-	private RubroGeneralDto rubroGeneralDto;
-	
-	
-	private List<DetallePedidoDto> detallePedidoDto = new ArrayList<DetallePedidoDto>();
-	
-
-	private List<DetalleFacturaDto> detalleFacturaDto = new ArrayList<DetalleFacturaDto>();
-    
-	
-	private List<ArticuloManufacturadoDetalleDto> articuloManufacturadoDetalleDto = new ArrayList<ArticuloManufacturadoDetalleDto>();
-		
-	//Constructores
-	
-	public ArticuloManufacturadoDto() {
-		//super();
-	}
-
-
-	public ArticuloManufacturadoDto(long id, int tiempoEstimadoCocina, String denominacion, double precioVenta, String imagen) {
-		super();
-		this.id = id;
-		this.tiempoEstimadoCocina = tiempoEstimadoCocina;
-		this.denominacion = denominacion;
-		this.precioVenta = precioVenta;
-		this.imagen = imagen;
-		
-	}
-
-	
-	//GETTERS Y SETTERS ------------------------------------------------------------------------
-
-	public long getId() {
-		return id;
-	}
-	
-	public void setId(long id) {
-		this.id = id;
-	}
-	
-	public int getTiempoEstimadoCocina() {
-		return tiempoEstimadoCocina;
-	}
-
-
-	public void setTiempoEstimadoCocina(int tiempoEstimadoCocina) {
-		this.tiempoEstimadoCocina = tiempoEstimadoCocina;
-	}
-
-
-	public String getDenominacion() {
-		return denominacion;
-	}
-
-
-	public void setDenominacion(String denominacion) {
-		this.denominacion = denominacion;
-	}
-
-
-	public double getPrecioVenta() {
-		return precioVenta;
-	}
-
-
-	public void setPrecioVenta(double precioVenta) {
-		this.precioVenta = precioVenta;
-	}
-
-
-	/*public RubroGeneral getRubroGeneral() {
-		return rubroGeneral;
-	}
-
-
-	public void setRubroGeneral(RubroGeneral rubroGeneral) {
-		this.rubroGeneral = rubroGeneral;
-	}
-	*/
-
-	public String getImagen() {
-		return imagen;
-	}
-
-	public void setImagen(String imagen) {
-		this.imagen = imagen;
-	}
-	
-	
-}
