--- src/main/java/com/example/demo/repository/ArticuloManufacturadoDetalleRepository.java
@@ -1,14 +1,0 @@
-package com.example.demo.repository;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.example.demo.entity.ArticuloManufacturadoDetalle;
-
-@Repository
-public interface ArticuloManufacturadoDetalleRepository extends JpaRepository<ArticuloManufacturadoDetalle, Long> {
-
-	public Page <ArticuloManufacturadoDetalle> findAll(Pageable pageable);
-}
