--- src/main/java/com/example/demo/repository/ArticuloManufacturadoRepository.java
@@ -1,15 +1,0 @@
-package com.example.demo.repository;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.example.demo.entity.ArticuloManufacturado;
-
-
-@Repository
-public interface ArticuloManufacturadoRepository extends JpaRepository<ArticuloManufacturado, Long> {
-
-	public Page <ArticuloManufacturado> findAll(Pageable pageable);
-}
